Complete Extended Opcodes for 0x8000
The last of the opcodes for 0x8000 opcodes were completed and tested
manually, using very masic observations. Full testing of the opcodes
will be needed once all opcodes are implemented.

Chip8_Stack debug messages where updated to display the index values
are pushed/ popped to as an int not a char, this way they show the
shack index correctly.

+++ Chip8Emulator/Chip8_Stack.cpp
@@ -8,7 +8,7 @@ void Chip8_Stack::Push(Chip8_Registers* registers, unsigned short value)
 
     if (DEBUG_MODE == 1)
     {
-        std::cout << "DEBUG: Pushing " << value << " to stack index " << registers->GetSPRegister() << std::endl;
+        std::cout << "DEBUG: Pushing " << value << " to stack index " << (int)registers->GetSPRegister() << std::endl;
     }
 
     stack[registers->GetSPRegister()] = value;
@@ -25,7 +25,7 @@ unsigned short Chip8_Stack::Pop(Chip8_Registers* registers)
 
     if (DEBUG_MODE == 1)
     {
-        std::cout << "DEBUG: Popping value from stack index " << registers->GetSPRegister() << std::endl;
+        std::cout << "DEBUG: Popping value from stack index " << (int)registers->GetSPRegister() << std::endl;
     }
 
     return stack[registers->GetSPRegister()];
